更新：完成序列化库基础功能

+++ Source/EasyFramework.Serialization/EasyFramework.Serialization/Serializers/FloatSerializer.cs
@@ -1,9 +1,9 @@
 namespace EasyFramework.Serialization
 {
     [EasySerializerPriority(EasySerializerProiority.System)]
-    public class IntSerializer : EasySerializerBase<int>
+    public class FloatSerializer : EasySerializerBase<float>
     {
-        protected override void Process(IArchive archive, ref int value)
+        protected override void Process(IArchive archive, ref float value)
         {
             archive.Process(ref value);
         }
